+++ app/src/main/java/com/xiaojianjun/wanandroid/ui/history/HistoryActivity.kt
@@ -87,7 +87,7 @@ class HistoryActivity : BaseVmActivity<HistoryViewModel>() {
         Bus.observe<Boolean>(USER_LOGIN_STATE_CHANGED, this, {
             mViewModel.updateListCollectState()
         })
-        Bus.observe<Pair<Int, Boolean>>(USER_COLLECT_UPDATED, this, {
+        Bus.observe<Pair<Long, Boolean>>(USER_COLLECT_UPDATED, this, {
             mViewModel.updateItemCollectState(it)
         })
     }
