+++ app/src/main/java/com/xiaojianjun/wanandroid/ui/main/home/latest/LatestFragment.kt
@@ -88,7 +88,7 @@ class LatestFragment : BaseVmFragment<LatestViewModel>(),
         Bus.observe<Boolean>(USER_LOGIN_STATE_CHANGED, viewLifecycleOwner, {
             mViewModel.updateListCollectState()
         })
-        Bus.observe<Pair<Int, Boolean>>(USER_COLLECT_UPDATED, viewLifecycleOwner, {
+        Bus.observe<Pair<Long, Boolean>>(USER_COLLECT_UPDATED, viewLifecycleOwner, {
             mViewModel.updateItemCollectState(it)
         })
     }
