+++ app/src/main/java/com/xiaojianjun/wanandroid/ui/main/home/popular/PopularFragment.kt
@@ -89,7 +89,7 @@ class PopularFragment : BaseVmFragment<PopularViewModel>(),
         Bus.observe<Boolean>(USER_LOGIN_STATE_CHANGED, viewLifecycleOwner, {
             mViewModel.updateListCollectState()
         })
-        Bus.observe<Pair<Int, Boolean>>(USER_COLLECT_UPDATED, viewLifecycleOwner, {
+        Bus.observe<Pair<Long, Boolean>>(USER_COLLECT_UPDATED, viewLifecycleOwner, {
             mViewModel.updateItemCollectState(it)
         })
     }
