+++ app/src/main/java/com/xiaojianjun/wanandroid/ui/main/home/project/ProjectFragment.kt
@@ -111,7 +111,7 @@ class ProjectFragment : BaseVmFragment<ProjectViewModel>(),
         Bus.observe<Boolean>(USER_LOGIN_STATE_CHANGED, viewLifecycleOwner, {
             mViewModel.updateListCollectState()
         })
-        Bus.observe<Pair<Int, Boolean>>(USER_COLLECT_UPDATED, viewLifecycleOwner, {
+        Bus.observe<Pair<Long, Boolean>>(USER_COLLECT_UPDATED, viewLifecycleOwner, {
             mViewModel.updateItemCollectState(it)
         })
     }
