+++ app/src/main/java/com/xiaojianjun/wanandroid/ui/search/result/SearchResultFragment.kt
@@ -93,7 +93,7 @@ class SearchResultFragment : BaseVmFragment<SearchResultViewModel>() {
         Bus.observe<Boolean>(USER_LOGIN_STATE_CHANGED, viewLifecycleOwner, {
             mViewModel.updateListCollectState()
         })
-        Bus.observe<Pair<Int, Boolean>>(USER_COLLECT_UPDATED, viewLifecycleOwner, {
+        Bus.observe<Pair<Long, Boolean>>(USER_COLLECT_UPDATED, viewLifecycleOwner, {
             mViewModel.updateItemCollectState(it)
         })
     }
